take the sync word in the 16-bit sx126x representation
Per review: the 16-bit form the sx126x writes to its sync word registers
becomes the canonical representation, since that is where new chips are
headed. The legacy single-byte form 0xYZ maps to 0xY4Z4 (public 0x34 is
0x3444, private 0x12 is 0x1424).

RadioKind::init_lora and set_lora_sync_word now take u16, and the cached
sync word is u16 so a full 16-bit value set on an sx126x survives a
re-init. The sx126x writes the two bytes through unchanged, which also
makes sync words with no single-byte equivalent reachable. The sx127x
and lr1110 map down to their single-byte registers and return the new
RadioError::InvalidSyncWord for values outside the 0xY4Z4 shape.

LoRa::new and with_syncword keep their signatures; with_syncword still
takes the legacy byte and converts on entry, so existing construction
behavior is unchanged.

+++ lora-phy/src/mod_traits.rs
@@ -34,9 +34,13 @@ pub enum IrqState {
 #[allow(async_fn_in_trait)]
 pub trait RadioKind {
     /// Initialize lora radio
-    async fn init_lora(&mut self, sync_word: u8) -> Result<(), RadioError>;
-    /// Apply a new LoRa sync word to the chip
-    async fn set_lora_sync_word(&mut self, sync_word: u8) -> Result<(), RadioError>;
+    ///
+    /// The sync word is given in the 16-bit sx126x register form; the legacy
+    /// single-byte form 0xYZ corresponds to 0xY4Z4.
+    async fn init_lora(&mut self, sync_word: u16) -> Result<(), RadioError>;
+    /// Apply a new LoRa sync word to the chip, in the same 16-bit form as
+    /// [`RadioKind::init_lora`]
+    async fn set_lora_sync_word(&mut self, sync_word: u16) -> Result<(), RadioError>;
     /// Create modulation parameters specific to the LoRa chip kind and type
     fn create_modulation_params(
         &self,
